Refactor websocket logic: move functions to a new file and update imports
